Refactoring 1.

+++ skiko/src/awtMain/kotlin/org/jetbrains/skiko/HardwareLayer.kt
@@ -66,7 +66,8 @@ internal open class HardwareLayer(
 
     val windowHandle: Long
         get() = useDrawingSurfacePlatformInfo(::getWindowHandle)
-    val currentDPI: Long
+
+    val currentDPI: Int
         get() = useDrawingSurfacePlatformInfo(::getCurrentDPI)
 
     val contentScale: Float
@@ -82,7 +83,7 @@ internal open class HardwareLayer(
 
     private external fun getContentHandle(platformInfo: Long): Long
     private external fun getWindowHandle(platformInfo: Long): Long
-    private external fun getCurrentDPI(platformInfo: Long): Long
+    private external fun getCurrentDPI(platformInfo: Long): Int
 
     private val _externalAccessible = externalAccessibleFactory?.invoke(this)
     private var _focusedAccessible: Accessible? = null
